Export table from wasm module by default
Previously in non-standalone mode we would create the table in
JS and then export it to the wasm module.  Now we always have the
wasm module create and export its own table, exceptin the dynamic
linking case way we do want the JS to be control.

This is a win for code size as well as consistency with standalone
mode.   It also makes life easier for those wishing to load
emscripten-generated modules with less JS glue since its once
less thing to export to the module.